+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add entrainment profiles to Betts-Miller convection, wet and dry [#976](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/976)
+
 ## v0.18
 
 - Adds a drag term to winds above a certain speed limit [#987](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/987/commits)